requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
pymongo==4.6.1
python-dotenv==1.0.1
bson==0.5.10
dnspython==2.6.1 